pyproject.toml
[tool.poetry]
name = "reading-sessions"
version = "0.1.0"
description = ""
authors = ["ASP <andrewpeshekhonov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
drf-yasg = "^1.21.7"

[tool.poetry.group.dev.dependencies]
faker = "^20.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
